feat: add installable executor companion plugin
