+++ .gitignore
@@ -4,3 +4,7 @@
 # Python bytecode generated by local verification
 __pycache__/
 *.py[cod]
+
+# Python packaging artifacts generated by local companion verification
+/build/
+/*.egg-info/
