+++ one_c_harness/release.json
@@ -1,0 +1,1 @@
+{"artifactId":"sha256:035f7262a026447429c5a83b688c26ac616c888225066fc5d5628decb2baecd4","schemaVersion":1}
