+++ scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility entrypoints for the canonical :mod:`one_c_harness` package."""
